Refactor `Load` trait to be independent of `Default`
A tuple could be formed by all `Load` types, but at least one
non-`Default` type. As a result, the tuple would not be `Load`,
while it should. Specialization is impossible due to the
blanket implementation of `Load` for `Writable` + `Default`.
This removes that implementation, shifting the responsibility
to `impl Load` to procedural macros down the line.

Still unimplemented: `Load` for `[Item; N]`.