Add `source()` to `Error` trait
The `error!`-generated implementation of `source()` loops through
all the symbols in the error backtrace, until one is found that
matches the call site of the `error!` macro. Since this is where
the `new()` method resides (and, consequently, where the backtrace
is generated) the immediately outer symbol represents the place
where `new()` is called.